requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python>=4.8.0
pyzbar>=0.1.9
tqdm
numpy
